auth-service/requirements.txt
flask==3.0.0
PyJWT==2.8.0
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0

